refactor(main): splash 抽到独立 main/splash.ts
createSplash + resolveSplashLogo（自包含、不耦合 index 模块状态）从 640 行的
main/index.ts 抽到 main/splash.ts，index 只 import createSplash。启动流程与
界面不变。
